refactor : 에러 처리용 메세지 객체 생성
