+++ src/main/java/roomescape/exception/ErrorResponse.java
@@ -1,0 +1,4 @@
+package roomescape.exception;
+
+public record ErrorResponse (String errorMessage){
+}
